+++ HISTORY.md
@@ -40,6 +40,7 @@
 * adjusted Dutch translation for variadic comparison ops
 * adjusted Catalan translation for variadic comparison ops
 * adjusted Brazilian Portuguese translation for variadic comparison ops
+* lists: tweaked distribution()
 
 ### 2023-02-26
 * threads: fixed #3183 - determining the length of huge strings fails